+++ prusti-viper/src/encoder/mir/contracts/mod.rs
@@ -1,0 +1,8 @@
+//! Procedure contracts enriched with borrow information.
+
+mod borrows;
+mod interface;
+
+pub(crate) use interface::{
+    ContractsEncoderInterface, ContractsEncoderState, ProcedureContract, ProcedureContractMirDef,
+};
