+++ prusti-viper/src/encoder/mir/procedures/encoder/mod.rs
@@ -1,11 +1,11 @@
 use self::{initialisation::InitializationData, specification_blocks::SpecificationBlocks};
 use super::MirProcedureEncoderInterface;
 use crate::encoder::{
-    borrows::ProcedureContractMirDef,
     errors::{ErrorCtxt, SpannedEncodingError, SpannedEncodingResult, WithSpan},
     mir::{
         casts::CastsEncoderInterface,
         constants::ConstantsEncoderInterface,
+        contracts::{ContractsEncoderInterface, ProcedureContractMirDef},
         errors::ErrorInterface,
         generics::MirGenericsEncoderInterface,
         panics::MirPanicsEncoderInterface,
