+++ prusti-viper/src/encoder/spec_function_encoder.rs
@@ -1,9 +1,12 @@
 use crate::encoder::snapshot::interface::SnapshotEncoderInterface;
-use crate::encoder::{Encoder, borrows::ProcedureContract};
+use crate::encoder::Encoder;
 use crate::encoder::errors::{ErrorCtxt, SpannedEncodingResult, WithSpan};
-use crate::encoder::borrows::compute_procedure_contract;
 use crate::encoder::mir_encoder::{MirEncoder, PlaceEncoder};
 use crate::encoder::mir::{
+    contracts::{
+        ContractsEncoderInterface,
+        ProcedureContract,
+    },
     pure::SpecificationEncoderInterface,
     specifications::SpecificationsInterface,
 };
@@ -56,18 +59,12 @@ impl<'p, 'v: 'p, 'tcx: 'v> SpecFunctionEncoder<'p, 'v, 'tcx> {
         let _post_name = self.encoder.encode_spec_func_name(self.procedure.get_id(),
                                                            SpecFunctionKind::Post);
 
-        let specs = self.encoder.get_procedure_specs(self.proc_def_id,self.substs);
-        if specs.is_none() {
+        if self.encoder.get_procedure_specs(self.proc_def_id, self.substs).is_none() {
             return Ok(vec![]);
         }
-        let specs = specs.unwrap();
 
-        let contract = compute_procedure_contract(
-            self.proc_def_id,
-            self.encoder.env(),
-            specs,
-            self.substs,
-        ).with_span(self.span)?.to_def_site_contract();
+        let contract = self.encoder.get_procedure_contract_for_def(self.proc_def_id, self.substs)
+            .with_span(self.span)?;
 
         let pre_func = self.encode_pre_spec_func(&contract)?;
         let post_func = self.encode_post_spec_func(&contract)?;
